+++ src/ui/ServerAdministration/index.tsx
@@ -34,6 +34,11 @@ import {
   wait,
 } from '../../utils';
 import {
+  ISyncedRealmToLoad,
+  RealmLoadingMode,
+  RealmToLoad,
+} from '../../utils/realms';
+import {
   IMenuGenerator,
   IMenuGeneratorProps,
 } from '../../windows/MenuGenerator';
@@ -167,9 +172,9 @@ class ServerAdministrationContainer
       this.setState({ isRealmOpening: true });
       try {
         // Let the UI update before sync waiting on the window to appear
-        const realm: ros.realms.ISyncedRealmToLoad = {
+        const realm: ISyncedRealmToLoad = {
           user: this.props.user,
-          mode: ros.realms.RealmLoadingMode.Synced,
+          mode: RealmLoadingMode.Synced,
           path,
           validateCertificates: this.props.validateCertificates,
         };
@@ -286,7 +291,7 @@ class ServerAdministrationContainer
       });
       await this.loadRealm({
         user: user.serialize(),
-        mode: ros.realms.RealmLoadingMode.Synced,
+        mode: RealmLoadingMode.Synced,
         path: '__admin',
         validateCertificates: this.props.validateCertificates,
       });
@@ -300,9 +305,7 @@ class ServerAdministrationContainer
     }
   }
 
-  protected async loadRealm(
-    realm: ros.realms.ISyncedRealmToLoad | ros.realms.ILocalRealmToLoad,
-  ) {
+  protected async loadRealm(realm: RealmToLoad) {
     if (
       this.certificateWasRejected &&
       realm.mode === 'synced' &&
@@ -359,27 +362,23 @@ class ServerAdministrationContainer
       const schema = dataImporter.generateSchema(format, paths);
       // Passing the generated schema to createRealm
       const newRealmFile = await this.createRealm(schema);
-      // Import the data into the Realm
-      const importer = dataImporter.getDataImporter(format, paths, schema);
       if (!this.state.user) {
         throw new Error('Cannot open realm without a user');
       }
-      const newRealm = await ros.realms.open({
-        user: this.state.user,
-        realmPath: newRealmFile.path,
-        ssl: { validateCertificates: this.props.validateCertificates },
-      });
-      // Import the data
-      importer.importInto(newRealm);
       // Open the Realm browser in "import mode"
-      const realm: ros.realms.ISyncedRealmToLoad = {
-        user: this.props.user,
-        mode: ros.realms.RealmLoadingMode.Synced,
-        path: newRealmFile.path,
-        validateCertificates: this.props.validateCertificates,
-      };
-      // Open the newly created realm
-      await main.showRealmBrowser({ realm });
+      await main.showRealmBrowser({
+        realm: {
+          user: this.props.user,
+          mode: RealmLoadingMode.Synced,
+          path: newRealmFile.path,
+          validateCertificates: this.props.validateCertificates,
+        },
+        import: {
+          format,
+          paths,
+          schema,
+        },
+      });
     } catch (err) {
       if (err.message === 'Realm creation cancelled') {
         // This is an expected expression to be thrown - no need to show it